First commit for my resume site.
This is my initial commit for my resume site.
PROBLEM: on C9, I can't see my css - it's showing my HTML as if I don't have a style sheet at all
I'm not sure if this is an issue with C9 or with my actual code. I am going to publish to test.